Tidy up py for final revision
